rename model with entity
