+++ as2_simulation_assets/as2_gazebo_assets/CMakeLists.txt
@@ -156,9 +156,9 @@ ament_target_dependencies(acro_bridge
   ${GZ_TRANSPORT}
 )
 
-add_executable(set_model_pose_bridge src/set_model_pose_bridge/set_model_pose_bridge.cpp
-src/set_model_pose_bridge/set_model_pose_bridge_node.cpp)
-ament_target_dependencies(set_model_pose_bridge
+add_executable(set_entity_pose_bridge src/set_entity_pose_bridge/set_entity_pose_bridge.cpp
+src/set_entity_pose_bridge/set_entity_pose_bridge_node.cpp)
+ament_target_dependencies(set_entity_pose_bridge
   rclcpp
   ros_gz_bridge
   std_msgs
@@ -176,7 +176,7 @@ install(TARGETS
   object_tf_broadcaster
   gimbal_bridge
   acro_bridge
-  set_model_pose_bridge
+  set_entity_pose_bridge
   DESTINATION lib/${PROJECT_NAME})
 
 # TODO(miferco97): Add the plugins to the CMakeLists.txt
