+++ as2_simulation_assets/as2_gazebo_assets/src/set_entity_pose_bridge/set_entity_pose_bridge.cpp
@@ -28,46 +28,16 @@
 
 
 /*!*******************************************************************************************
- *  \file       set_model_pose_bridge.cpp
- *  \brief      Set model pose bridge implementation file.
+ *  \file       set_entity_pose_bridge.cpp
+ *  \brief      Set entity pose bridge implementation file.
  *  \authors    Javier Melero Deza
  *              Pedro Arias Pérez
- *              Rafael Pérez Seguí
- *              Miguel Fernández Cortizas
- *              David Pérez Saura
- *
- *  \copyright  Copyright (c) 2022 Universidad Politécnica de Madrid
- *              All Rights Reserved
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- * 3. Neither the name of the copyright holder nor the names of its contributors
- *    may be used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
- * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ********************************************************************************/
 
-#include "as2_gazebo_assets/set_model_pose_bridge.hpp"
+#include "as2_gazebo_assets/set_entity_pose_bridge.hpp"
 
-SetModelPoseBridge::SetModelPoseBridge()
-: Node("set_model_pose_bridge")
+SetEntityPoseBridge::SetEntityPoseBridge()
+: Node("set_entity_pose_bridge")
 {
   this->declare_parameter<std::string>("world_name");
   this->get_parameter("world_name", world_name_);
@@ -75,7 +45,7 @@ SetModelPoseBridge::SetModelPoseBridge()
   ps_srv_sub_ = this->create_service<as2_msgs::srv::SetPoseWithID>(
     "/world/" + world_name_ + "/set_pose",
     std::bind(
-      &SetModelPoseBridge::setModelPoseServiceCallback, this, std::placeholders::_1,
+      &SetEntityPoseBridge::setEntityPoseServiceCallback, this, std::placeholders::_1,
       std::placeholders::_2));
 
   // Initialize the gz node
@@ -83,7 +53,7 @@ SetModelPoseBridge::SetModelPoseBridge()
   std::string set_model_pose_service = "/world/" + world_name_ + "/set_pose";
 }
 
-void SetModelPoseBridge::setModelPoseServiceCallback(
+void SetEntityPoseBridge::setEntityPoseServiceCallback(
   const as2_msgs::srv::SetPoseWithID::Request::SharedPtr request,
   as2_msgs::srv::SetPoseWithID::Response::SharedPtr result)
 {
